// packages/solutions-components/src/utils/interfaces.ts
export enum EUpdateType {
  Add,
  Update,
  Remove,
  None,
  Obsolete
}
// ...
export interface IResourcePath extends IDeployFileCopyPath {
  blob?: any; // This will only be set when a new file is uploaded (add or update)
  sourceFileName?: string; // This will only be set when a file is being updated
  updateType: EUpdateType;
}